Code style changes
Automatic commit - 834f50e1e2fb872db27a1a686923f8ff39d17863

+++ src/BillaBear/Pricing/PriceCalculation.php
@@ -14,7 +14,7 @@ readonly class PriceCalculation
 {
     public function __construct(
         public Money $money,
-        public int|float $quantity,
+        public float|int $quantity,
         public Money $netPrice,
     ) {
     }
